+++ src/context/directory/handlers/selfServiceProfiles.ts
@@ -39,7 +39,7 @@ function parse(context: DirectoryContext): ParsedSelfServiceProfiles {
 }
 
 async function dump(context: DirectoryContext): Promise<void> {
-  const { selfServiceProfiles } = context.assets;
+  const { selfServiceProfiles, userAttributeProfilesWithId } = context.assets;
   if (!selfServiceProfiles) return;
 
   const selfServiceProfilesFolder = path.join(
@@ -58,6 +58,17 @@ async function dump(context: DirectoryContext): Promise<void> {
 
     if ('updated_at' in profile) {
       delete profile.updated_at;
+    }
+
+    if (profile.user_attribute_profile_id) {
+      const p = userAttributeProfilesWithId?.find(
+        (uap) => uap.id === profile.user_attribute_profile_id
+      );
+      profile.user_attribute_profile_id = p?.name || profile.user_attribute_profile_id;
+
+      if (profile.user_attributes?.length === 0) {
+        delete profile.user_attributes;
+      }
     }
 
     dumpJSON(ssProfileFile, profile);