+++ test/context/yaml/selfServiceProfiles.test.js
@@ -196,4 +196,70 @@ describe('#YAML context selfServiceProfiles', () => {
     const dumped = await handler.dump(context);
     expect(dumped).to.deep.equal({ selfServiceProfiles: target });
   });
+
+  it('should dump selfServiceProfiles and convert user_attribute_profile_id to names on dump', async () => {
+    const dir = path.join(testDataDir, 'yaml', 'selfServiceProfilesDump');
+    cleanThenMkdir(dir);
+    const context = new Context(
+      { AUTH0_INPUT_FILE: path.join(dir, './test.yml') },
+      mockMgmtClient()
+    );
+    const selfServiceProfiles = [
+      {
+        name: 'test-self-service-profile',
+        description: 'test Self-Service Profile',
+        allowed_strategies: [
+          'adfs',
+          'google-apps',
+          'keycloak-samlp',
+          'oidc',
+          'okta',
+          'samlp',
+          'waad',
+        ],
+        branding: { colors: { primary: '#19aecc' } },
+        customText: {
+          en: {
+            'get-started': {
+              introduction:
+                "Welcome! With <p>only a few steps</p> you'll be able to setup your new connection.",
+            },
+          },
+        },
+        user_attribute_profile_id: 'uap_12345',
+      },
+    ];
+
+    const target = [
+      {
+        name: 'test-self-service-profile',
+        description: 'test Self-Service Profile',
+        allowed_strategies: [
+          'adfs',
+          'google-apps',
+          'keycloak-samlp',
+          'oidc',
+          'okta',
+          'samlp',
+          'waad',
+        ],
+        branding: { colors: { primary: '#19aecc' } },
+        customText: {
+          en: {
+            'get-started': {
+              introduction:
+                "Welcome! With <p>only a few steps</p> you'll be able to setup your new connection.",
+            },
+          },
+        },
+        user_attribute_profile_id: 'profile1',
+      },
+    ];
+
+    context.assets.selfServiceProfiles = selfServiceProfiles;
+    context.assets.userAttributeProfiles = [{ id: 'uap_12345', name: 'profile1' }];
+
+    const dumped = await handler.dump(context);
+    expect(dumped).to.deep.equal({ selfServiceProfiles: target });
+  });
 });